+++ app/src/main/java/com/example/robotpatrolapp/model/RobotPosition.kt
@@ -1,7 +1,7 @@
 package com.example.robotpatrolapp.model
 
 data class RobotPosition(
-    var x: Float,  // X coordinate
-    var y: Float,  // Y coordinate
-    var theta: Float  // Orientation in degrees (0° = facing right, 90° = facing up)
+    var x: Double,  // X coordinate
+    var y: Double,  // Y coordinate
+    var theta: Double  // Orientation in degrees (0° = facing right, 90° = facing up)
 )
